requirements.txt
django
django-storages[azure]
pillow
python-decouple
python-dotenv
opencv-python-headless
openai